Define size_t and ssize_t in terms of uintptr_t and intptr_t, since the former and the latter should be the same width, as both refer to objects in memory.
XXX Need to make sure to not use size_t for non-memory objects.  off_t?

+++ mk/core/types.h
@@ -11,8 +11,8 @@
 
 	/* Size types and friends.  */
 
-typedef	uint64_t	size_t;
-typedef	int64_t		ssize_t;
+typedef	uintptr_t	size_t;
+typedef	intptr_t	ssize_t;
 
 typedef	int64_t		off_t;
